Updated files to create a selective task stream
Created new column "TaskField" in the table "Tasks" to store the field
of study for the task in question. Altered index.php to display only
tasks that correspond to the field of study of the user. Made small
changes in the other files to accommodate the implementation of the task
stream